requirements.txt
requests>=2.31.0
PyYAML>=6.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
